refactor: refactoring

+++ algorithms/heapsort.c
@@ -1,8 +1,9 @@
 #include "../helpers/sort_helpers.h"
+#include "algorithms.h"
 
 #define largest(a, length, i, j, k)                    \
     ({                                                 \
-        unsigned int largest = i;                      \
+        size_t largest = i;                      \
         if (j < length && cmp(a + j, a + largest) > 0) \
         {                                              \
             largest = j;                               \
@@ -14,9 +15,9 @@
         largest;                                       \
     })
 
-static void heapify(int a[], unsigned int length, unsigned int i)
+static void heapify(int a[], size_t length, size_t i)
 {
-    unsigned int j, largest;
+    size_t j, largest;
     do
     {
         j = (i << 1) + 1;
@@ -24,15 +25,15 @@ static void heapify(int a[], unsigned int length, unsigned int i)
     } while (largest != i && (i = swap(a, largest, i), 1));
 }
 
-void heapsort(int a[], unsigned int length)
+void heapsort(int a[], size_t length)
 {
     counters.heapsort_counter++;
-    for (unsigned int i = (length >> 1) - 1; i < U32_MAX; i--)
+    for (size_t i = (length >> 1) - 1; i < U32_MAX; i--)
     {
         heapify(a, length, i);
     }
 
-    for (unsigned int i = length - 1; i < U32_MAX; i--)
+    for (size_t i = length - 1; i < U32_MAX; i--)
     {
         swap(a, 0, i);
         heapify(a, i, 0);
