+++ algorithms/insertion_sort.c
@@ -1,8 +1,9 @@
 #include "../helpers/sort_helpers.h"
+#include "algorithms.h"
 
-static void insert(int a[], unsigned int i)
+static void insert(int a[], size_t i)
 {
-    unsigned int j = i;
+    size_t j = i;
     while (j > 0 && cmp(a + j - 1, a + j) > 0)
     {
         swap(a, j, j - 1);
@@ -10,10 +11,10 @@ static void insert(int a[], unsigned int i)
     }
 }
 
-void insertion_sort(int a[], unsigned int length)
+void insertion_sort(int a[], size_t length)
 {
     counters.insertion_sort_counter++;
-    for (unsigned int i = 1; i < length; i++)
+    for (size_t i = 1; i < length; i++)
     {
         insert(a, i);
     }
