+++ algorithms/introsort.c
@@ -1,12 +1,13 @@
 #include <math.h>
 
 #include "../helpers/sort_helpers.h"
+#include "algorithms.h"
 
-static void introsort_rec(int a[], unsigned int left, unsigned int right, unsigned int depth_limit)
+static void introsort_rec(int a[], size_t left, size_t right, size_t depth_limit)
 {
     counters.recursion_counter++;
 
-    unsigned int length = right - left + 1;
+    size_t length = right - left + 1;
 
     if (length <= MAX_THRESH)
     {
@@ -20,13 +21,13 @@ static void introsort_rec(int a[], unsigned int left, unsigned int right, unsign
         return;
     }
 
-    unsigned int pivot = partition(a, left, right, MID, NULL);
+    size_t pivot = partition(a, left, right, MID, NULL);
     introsort_rec(a, left, pivot, depth_limit - 1);
     introsort_rec(a, pivot + 1, right, depth_limit - 1);
 }
 
-void introsort(int a[], unsigned int length)
+void introsort(int a[], size_t length)
 {
-    unsigned int depth_limit = 2 * log(length);
+    size_t depth_limit = 2 * log(length);
     introsort_rec(a, 0, !length ? 0 : length - 1, depth_limit);
 }
