+++ algorithms/quicksort_std.c
@@ -1,8 +1,9 @@
 #include <stdlib.h>
 
 #include "../helpers/sort_helpers.h"
+#include "algorithms.h"
 
-void quicksort_std(int a[], unsigned int length)
+void quicksort_std(int a[], size_t length)
 {
     qsort(a, length, sizeof(int), cmp);
 }
