+++ helpers/array_helpers.c
@@ -5,13 +5,13 @@
 #include "array_helpers.h"
 #include "sort_helpers.h"
 
-unsigned long int array_from_stdin(int *a[])
+size_t array_from_stdin(int *a[])
 {
-    long int tmp_size;
-    unsigned long int size;
+    int tmp_size;
+    size_t size;
 
     // Read array size
-    if (scanf("%ld", &tmp_size) != 1)
+    if (scanf("%d", &tmp_size) != 1)
     {
         fprintf(stderr, "Error reading size from input.\n");
         exit(EXIT_FAILURE);
@@ -20,11 +20,11 @@ unsigned long int array_from_stdin(int *a[])
     // Check if size is valid
     if (tmp_size < 0)
     {
-        fprintf(stderr, "Array size must be a positive integer, got %ld.\n", tmp_size);
+        fprintf(stderr, "Array size must be a positive integer, got %d.\n", tmp_size);
         exit(EXIT_FAILURE);
     }
 
-    size = (unsigned long int)tmp_size;
+    size = (size_t)tmp_size;
 
     // Allocate memory for array
     *a = (int *)malloc(size * sizeof(int));
@@ -35,7 +35,7 @@ unsigned long int array_from_stdin(int *a[])
     }
 
     // Read array elements
-    for (unsigned long int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
     {
         if (scanf("%d", &(*a)[i]) != 1)
         {
@@ -48,9 +48,9 @@ unsigned long int array_from_stdin(int *a[])
     return size;
 }
 
-void array_dump(int a[], unsigned int length)
+void array_dump(int a[], size_t length)
 {
-    for (unsigned int i = 0; i < length; i++)
+    for (size_t i = 0; i < length; i++)
     {
         printf("%i", a[i]);
         printf((i < length - 1) ? " " : "\n");
@@ -58,16 +58,16 @@ void array_dump(int a[], unsigned int length)
     printf("--------------------------------------------------------------------------------------------------------------------------------------------\n");
 }
 
-int *array_copy(int array[], unsigned int length)
+int *array_copy(int array[], size_t length)
 {
     int *copy = malloc(length * sizeof(int));
     memcpy(copy, array, length * sizeof(int));
     return copy;
 }
 
-bool array_is_sorted(int a[], unsigned int length)
+bool array_is_sorted(int a[], size_t length)
 {
-    unsigned int i = 1;
+    size_t i = 1;
     while (i < length && cmp(a + i - 1, a + i) <= 0)
     {
         i++;
@@ -76,10 +76,10 @@ bool array_is_sorted(int a[], unsigned int length)
     return !length || i == length;
 }
 
-static unsigned int array_value_count(int a[], unsigned int length, int value)
+static size_t array_value_count(int a[], size_t length, int value)
 {
-    unsigned int count = 0;
-    for (unsigned int i = 0; i < length; i++)
+    size_t count = 0;
+    for (size_t i = 0; i < length; i++)
     {
         if (a[i] == value)
         {
@@ -89,14 +89,14 @@ static unsigned int array_value_count(int a[], unsigned int length, int value)
     return count;
 }
 
-bool array_is_permutation_of(int a[], int b[], unsigned int length)
+bool array_is_permutation_of(int a[], int b[], size_t length)
 {
-    unsigned int i = 0;
+    size_t i = 0;
     bool result = true;
     while (i < length && result)
     {
-        unsigned int a_count = array_value_count(a, length, a[i]);
-        unsigned int b_count = array_value_count(b, length, a[i]);
+        size_t a_count = array_value_count(a, length, a[i]);
+        size_t b_count = array_value_count(b, length, a[i]);
         result = (a_count == b_count);
         i++;
     }
